requirements.txt
flask
openai-whisper
transformers
sentence-transformers
faiss-cpu
torch
